sesson 57

+++ Modules/Mlk/Advertising/Http/Controllers/AdvertisingController.php
@@ -4,6 +4,7 @@ namespace Mlk\Advertising\Http\Controllers;
 
 use App\Http\Controllers\Controller;
 use Mlk\Advertising\Http\Requests\AdvertisingRequest;
+use Mlk\Advertising\Models\Advertising;
 use Mlk\Advertising\Repositories\AdvertisingRepo;
 use Mlk\Advertising\Services\AdvertisingService;
 use Mlk\Share\Repositories\ShareRepo;
@@ -11,6 +12,8 @@ use Mlk\Share\Services\ShareService;
 
 class AdvertisingController extends Controller
 {
+    private string $class = Advertising::class;
+
     public AdvertisingRepo $repo;
     public AdvertisingService $service;
 
@@ -22,17 +25,22 @@ class AdvertisingController extends Controller
 
     public function index()
     {
+        $this->authorize('index', $this->class);
         $advertisings = $this->repo->index()->paginate(10);
+
         return view('Advs::index', compact('advertisings'));
     }
 
     public function create()
     {
+        $this->authorize('index', $this->class);
         return view('Advs::create');
     }
 
     public function store(AdvertisingRequest $request)
     {
+        $this->authorize('index', $this->class);
+
         [$imageName, $imagePath] = ShareService::uploadImage($request->file('image'), null);
         $this->service->store($request, $imagePath, $imageName);
 
@@ -42,17 +50,18 @@ class AdvertisingController extends Controller
 
     public function edit($id)
     {
+        $this->authorize('index', $this->class);
         $advertising = $this->repo->findById($id);
+
         return view('Advs::edit', compact('advertising'));
     }
 
     public function update(AdvertisingRequest $request, $id)
     {
-        $file = $request->file('image');
+        $this->authorize('index', $this->class);
+
         $advertising = $this->repo->findById($id);
-
-        [$imageName, $imagePath] = $this->uploadImage($file, $advertising);
-
+        [$imageName, $imagePath] = $this->uploadImage($request->file('image'), $advertising);
         $this->service->update($request, $id, $imagePath, $imageName);
 
         ShareRepo::successMessage('ویرایش تبلیغات');
@@ -61,6 +70,8 @@ class AdvertisingController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('index', $this->class);
+
         $article = $this->repo->findById($id);
         $this->service->deleteImage($article);
         $this->repo->delete($id);
